major modifications - iterator tools
