+++ CMakeLists.txt
@@ -105,8 +105,6 @@ ADD_LIBRARY(itertools INTERFACE)
 TARGET_INCLUDE_DIRECTORIES(itertools INTERFACE
 	$<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include/itertools>
 	$<INSTALL_INTERFACE:${CMAKE_CURRENT_BINARY_DIR}/lib>)
-ADD_EXECUTABLE(itertools-example ${PROJECT_SOURCE_DIR}/src/itertoolstest.c++)
-target_link_libraries(itertools-example itertools)
 
 ADD_LIBRARY(cvmclushelp SHARED ${PROJECT_SOURCE_DIR}/src/cvmclushelp.c++)
 ADD_LIBRARY(cvhelp SHARED ${PROJECT_SOURCE_DIR}/src/cvhelp.c++)
