+++ include/itertools/base.h
