+++ include/itertools/enumerate.h
@@ -1,116 +1,110 @@
-#ifndef __ITERTOOLS_ENUMERATE_HPP__
-#define __ITERTOOLS_ENUMERATE_HPP__
+#ifndef __HYSYS_ITERATORTOOLS_ENUMERATOR_HPP__
+#define __HYSYS_ITERATORTOOLS_ENUMERATOR_HPP__
 
-#include "itertoolsbase.h"
-#include "itertoolswrapper.h"
+#include "hysysiteratortools.hpp"
 
-namespace itertools {
-template <typename Index, typename Elem>
-using EnumBasePair = std::pair<Index, Elem>;
+namespace HysysIteratorTools
+	{
+    // Held by the Enumerable::Iterator.  Has a .index, and a
+    // .element referencing the value yielded by the subiterator
+    template <typename Index, typename Elem>
+    class EnumeratorDataHolder : public std::pair<Index, Elem> {
+        using std::pair<Index,Elem>::pair
 
-// "yielded" by the Enumerable::Iterator.  Has a .index, and a
-// .element referencing the value yielded by the subiterator
-template <typename Index, typename Elem>
-class EnumIteratorYield : public EnumBasePair<Index, Elem> {
-  using BasePair = EnumBasePair<Index, Elem>;
-  using BasePair::BasePair;
+    public:
+        typename std::pair<Index, Elem>::first_type index = std::pair<Index, Elem>::first;
+        typename std::pair<Index, Elem>::second_type element = std::pair<Index, Elem>::second;
+        };
 
- public:
-  typename BasePair::first_type index = BasePair::first;
-  typename BasePair::second_type element = BasePair::second;
-};
+    template <typename Container, typename Index>
+    class Enumerable;
 
-template <typename Container, typename Index>
-class Enumerable;
+    template <typename Container>
+    constexpr decltype(auto) enumerate = [](Container&& container)
+        {
+        return Enumerable<Container, std::size_t>(std::forward<Container>(container));
+        };
 
-using EnumerateFn = IterToolFnOptionalBindSecond<Enumerable, std::size_t>;
-constexpr EnumerateFn enumerate{};
-}  // namespace itertools
+    template <typename Index, typename Elem>
+    struct std::tuple_size<EnumeratorDataHolder<Index, Elem>> : public tuple_size<std::pair<Index, Elem>> {};
 
-namespace std {
-template <typename Index, typename Elem>
-struct tuple_size<itertools::EnumIteratorYield<Index, Elem>>
-    : public tuple_size<itertools::EnumBasePair<Index, Elem>> {};
-
-template <std::size_t N, typename Index, typename Elem>
-struct tuple_element<N, itertools::EnumIteratorYield<Index, Elem>>
-    : public tuple_element<N, itertools::EnumBasePair<Index, Elem>> {};
-}  // namespace std
+    template <std::size_t N, typename Index, typename Elem>
+    struct std::tuple_element<N, EnumeratorDataHolder<Index, Elem>> : public tuple_element<N, std::pair<Index, Elem>> {};
+	}
 
 template <typename Container, typename Index>
-class itertools::Enumerable {
- private:
-  Container _container;
-  const Index _start;
-  friend EnumerateFn;
-  // Value constructor for use only in the enumerate function
-  Enumerable(Container&& container, Index start)
-      : _container(std::forward<Container>(container)), _start{start} {}
+class HysysIteratorTools::Enumerable 
+    {
+    private:
+        Container container_;
+        const Index start_;
+        
+        friend EnumerateFn;
+        
+        // private Value constructor
+        Enumerable(Container&& container, Index start) 
+            : container_(std::forward<Container>(container)), start_{ start } {}
 
- public:
-  Enumerable(Enumerable&&) = default;
+    public:
+        Enumerable(Enumerable&&) = default;
 
-  template <typename T>
-  using IteratorYield = EnumIteratorYield<Index, iterator_deref<T>>;
+        template <typename T>
+        using Data = EnumeratorDataHolder<Index, iterator_deref<T>>;
 
-  //  Holds an iterator of the contained type and an Index for the
-  //  index_.  Each call to ++ increments both of these data members.
-  //  Each dereference returns an IterYield.
-  template <typename ContainerT>
-  class Iterator {
-   private:
-    template <typename>
-    friend class Iterator;
-    IteratorWrapper<ContainerT> _sub_iter;
-    Index _index;
+        //  Holds an iterator of the contained type and an Index for the
+        // index_.  Each call to ++ increments both of these data members.
+        // Each dereference returns an IterYield.
+        template <typename ContainerT>
+        class Iterator 
+            {
+            private:
+                
+                template <typename>
+                friend class Iterator;
+                
+                iterator_t<ContainerT> sub_iter_;
+                Index index_;
+            
+            public:
+                using iterator_category = std::input_iterator_tag;
+                using value_type = Data<ContainerT>
+                using difference_type = std::ptrdiff_t;
+                using pointer = value_type*;
+                using reference = value_type&;
+                
+                Iterator(iterator_t<ContainerT>&& sub_iter, Index start)
+                    : sub_iter_{ std::move(sub_iter) }, index_{ start } {}
+                
+                Data<ContainerT> operator*() { return { index_, *sub_iter_ }; }
+                
+                ArrowProxy<Data<ContainerT>> operator->() { return { **this }; }
+                
+                Iterator& operator++() 
+                    {
+                    ++sub_iter_;
+                    ++index_;
+                    return *this;
+                    }
+                
+                Iterator operator++(int) 
+                    {
+                    auto ret = *this;
+                    ++*this;
+                    return ret;
+                    }
+                
+                template <typename T>
+                bool operator!=(const Iterator<T>& other) const { return sub_iter_ != other.sub_iter_; }
+                
+                template <typename T>
+                bool operator==(const Iterator<T>& other) const { return !(*this != other); }
+            };
+        
+        Iterator<Container> begin() { return { fancy_getters::begin(container_), start_ }; }
+        Iterator<Container> end() { return { fancy_getters::end(container_), start_ }; }
+        
+        Iterator<make_const_t<Container>> begin() const { return { fancy_getters::begin(std::as_const(container_)), start_ }; }
+        Iterator<make_const_t<Container>> end() const { return { fancy_getters::end(std::as_const(container_)), start_ }; }
+    };
 
-   public:
-    using iterator_category = std::input_iterator_tag;
-    using value_type = IteratorYield<ContainerT>;
-    using difference_type = std::ptrdiff_t;
-    using pointer = value_type*;
-    using reference = value_type&;
-
-    Iterator(IteratorWrapper<ContainerT>&& sub_iter, Index start)
-	: _sub_iter{std::move(sub_iter)}, _index{start} {}
-
-    IteratorYield<ContainerT> operator*() { return {_index, *_sub_iter}; }
-
-    ArrowProxy<IteratorYield<ContainerT>> operator->() { return {**this}; }
-
-    Iterator& operator++() {
-      ++_sub_iter;
-      ++_index;
-      return *this;
-    }
-
-    Iterator operator++(int) {
-      auto ret = *this;
-      ++*this;
-      return ret;
-    }
-
-    template <typename T>
-    bool operator!=(const Iterator<T>& other) const {
-      return _sub_iter != other._sub_iter;
-    }
-
-    template <typename T>
-    bool operator==(const Iterator<T>& other) const {
-      return !(*this != other);
-    }
-  };
-
-  Iterator<Container> begin() { return {getters::begin(_container), _start}; }
-
-  Iterator<Container> end() { return {getters::end(_container), _start}; }
-
-  Iterator<AsConst<Container>> begin() const {
-    return {getters::begin(std::as_const(_container)), _start};
-  }
-
-  Iterator<AsConst<Container>> end() const {
-    return {getters::end(std::as_const(_container)), _start};
-  }
-};
-#endif	// __ITERTOOLS_ENUMERATE_HPP__
+#endif //__HYSYS_ITERATORTOOLS_ENUMERATOR_HPP__
