--- include/itertools/itertools.h
@@ -1,6 +1,0 @@
-#ifndef __ITERTOOLS_HPP__
-#define __ITERTOOLS_HPP__
-
-#include "range.h"
-
-#endif	//__ITERTOOLS_HPP__
